+++ src/hocuspocus/extensions/authentication/alkemio-authenticator/index.ts
@@ -1,0 +1,4 @@
+export * from './alkemio.authenticator';
+export * from './authorization.service';
+export * from './authentication.service';
+export * from './alkemio.authenticator.factory';
