+++ src/hocuspocus/extensions/authentication/authentication.module.ts
@@ -1,10 +1,22 @@
 import { Module } from '@nestjs/common';
 import { UtilModule } from '@src/services/util';
-import AuthenticationFactory from './authentication.factory';
+import { HocuspocusConnectionService } from '@src/hocuspocus/services';
+import {
+  AuthorizationService,
+  AlkemioAuthenticator,
+  AuthenticationService,
+  AlkemioAuthenticatorFactory,
+} from './alkemio-authenticator';
 
 @Module({
   imports: [UtilModule],
-  providers: [AuthenticationFactory],
-  exports: [AuthenticationFactory],
+  providers: [
+    AuthenticationService,
+    AuthorizationService,
+    HocuspocusConnectionService,
+    AlkemioAuthenticator,
+    AlkemioAuthenticatorFactory,
+  ],
+  exports: [AlkemioAuthenticatorFactory],
 })
 export class AuthenticationModule {}
