+++ src/hocuspocus/extensions/authentication/exceptions/authentication.exception.ts
