+++ src/hocuspocus/extensions/authentication/exceptions/forbidden.exception.ts
