+++ src/hocuspocus/extensions/authentication/exceptions/index.ts
@@ -1,0 +1,2 @@
+export * from './forbidden.exception';
+export * from './authentication.exception';
