+++ src/hocuspocus/extensions/authentication/types/index.ts
@@ -1,0 +1,2 @@
+export * from './auth.interfaces';
+export * from './read.only.code';
