+++ src/hocuspocus/extensions/authentication/types/read.only.code.ts
